backup/llm-chatbot-framework/requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
tqdm>=4.62.0

# LLM and embedding dependencies
openai>=1.0.0
dspy-ai>=2.0.0
transformers>=4.20.0
torch>=1.10.0
sentence-transformers>=2.2.0

# Text processing
nltk>=3.6.0
PyPDF2>=2.0.0
pdfplumber>=0.7.0

# Vector database
faiss-cpu>=1.7.0

# Utilities
python-dotenv>=0.19.0
requests>=2.26.0

# Document processing
python-docx>=0.8.11
ebooklib>=0.17.1
beautifulsoup4>=4.12.0

# Web application
Flask>=2.3.0
Flask-Cors>=4.0.0
Werkzeug>=2.3.0

# Utilities
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Fine-tuning dependencies
peft>=0.5.0  # Parameter-Efficient Fine-Tuning
datasets>=2.12.0  # HuggingFace datasets
accelerate>=0.20.0  # For faster training
bitsandbytes>=0.40.0  # For quantization
scikit-learn>=1.0.0  # For evaluation metrics
backoff>=2.2.0  # For retry logic
jsonlines>=3.1.0  # For JSONL file processing
langchain_openai>=0.0.1  # For OpenAI integration

# Evaluation metrics
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.1
nltk>=3.8.1
pyyaml>=6.0.1
numpy>=1.24.0 